# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para aplicación de finanzas personales",
        "title": "API Fintech",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "fintech-production-5841.up.railway.app",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/budgets": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo presupuesto mensual con asignaciones por categoría",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Crear presupuesto mensual",
                "parameters": [
                    {
                        "description": "Datos del presupuesto",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BudgetSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/budgets/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene el presupuesto del mes actual del usuario",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Obtener presupuesto actual",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BudgetSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/budgets/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene el dashboard con información del presupuesto actual, gastos y alertas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Obtener dashboard principal",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BudgetDashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/budgets/month": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene el presupuesto de un mes específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Obtener presupuesto por mes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Año (ej: 2024)",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Mes (1-12)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BudgetSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/budgets/rollover": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Procesa el rollover diario de saldos no gastados (endpoint interno/cron)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Procesar rollover diario",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/budgets/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un presupuesto existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "budgets"
                ],
                "summary": "Actualizar presupuesto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del presupuesto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de actualización",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BudgetSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AllocationAlert": {
            "type": "object",
            "properties": {
                "alert_type": {
                    "description": "\"warning\", \"danger\", \"over_budget\"",
                    "type": "string"
                },
                "allocated_amount": {
                    "type": "number"
                },
                "category_icon": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "progress_percent": {
                    "type": "number"
                },
                "spent_amount": {
                    "type": "number"
                }
            }
        },
        "dto.AllocationSummaryResponse": {
            "type": "object",
            "properties": {
                "alert_threshold": {
                    "type": "number"
                },
                "allocated_amount": {
                    "type": "number"
                },
                "allocation_percent": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/dto.CategorySummaryResponse"
                },
                "current_daily_limit": {
                    "type": "number"
                },
                "daily_limit": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_over_budget": {
                    "type": "boolean"
                },
                "progress_percent": {
                    "type": "number"
                },
                "remaining_amount": {
                    "type": "number"
                },
                "should_alert": {
                    "type": "boolean"
                },
                "spent_amount": {
                    "type": "number"
                }
            }
        },
        "dto.BudgetDashboardResponse": {
            "type": "object",
            "properties": {
                "category_alerts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AllocationAlert"
                    }
                },
                "current_budget": {
                    "$ref": "#/definitions/dto.BudgetSummaryResponse"
                },
                "month_total": {
                    "type": "number"
                },
                "quick_stats": {
                    "$ref": "#/definitions/dto.BudgetQuickStats"
                },
                "today_expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ExpenseSummaryResponse"
                    }
                },
                "today_total": {
                    "type": "number"
                },
                "week_total": {
                    "type": "number"
                }
            }
        },
        "dto.BudgetQuickStats": {
            "type": "object",
            "properties": {
                "average_daily_spent": {
                    "type": "number"
                },
                "categories_on_track": {
                    "type": "integer"
                },
                "categories_over_budget": {
                    "type": "integer"
                },
                "days_until_payday": {
                    "type": "integer"
                },
                "recommended_daily": {
                    "type": "number"
                },
                "total_categories": {
                    "type": "integer"
                }
            }
        },
        "dto.BudgetSummaryResponse": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AllocationSummaryResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_current_month": {
                    "type": "boolean"
                },
                "month": {
                    "type": "integer"
                },
                "period_string": {
                    "type": "string"
                },
                "progress_percent": {
                    "type": "number"
                },
                "remaining_amount": {
                    "type": "number"
                },
                "remaining_days": {
                    "type": "integer"
                },
                "spent_amount": {
                    "type": "number"
                },
                "total_amount": {
                    "type": "number"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CategorySummaryResponse": {
            "type": "object",
            "properties": {
                "can_be_deleted": {
                    "type": "boolean"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_default": {
                    "type": "boolean"
                },
                "is_user_category": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateAllocationRequest": {
            "type": "object",
            "required": [
                "allocated_amount",
                "category_id"
            ],
            "properties": {
                "alert_threshold": {
                    "description": "0.0 a 1.0",
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "allocated_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "category_id": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateBudgetRequest": {
            "type": "object",
            "required": [
                "allocations",
                "month",
                "total_amount",
                "year"
            ],
            "properties": {
                "allocations": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.CreateAllocationRequest"
                    }
                },
                "month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "total_amount": {
                    "type": "number"
                },
                "year": {
                    "type": "integer",
                    "maximum": 2030,
                    "minimum": 2020
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ExpenseSummaryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "can_be_cancelled": {
                    "type": "boolean"
                },
                "can_be_modified": {
                    "type": "boolean"
                },
                "category": {
                    "$ref": "#/definitions/dto.CategorySummaryResponse"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "description": "ISO format",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "formatted_amount": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "merchant": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/entity.ExpenseSource"
                },
                "status": {
                    "$ref": "#/definitions/entity.ExpenseStatus"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time_ago": {
                    "description": "\"Hace 2 horas\"",
                    "type": "string"
                },
                "triggered_alert": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateAllocationRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "alert_threshold": {
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "allocated_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "dto.UpdateBudgetRequest": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UpdateAllocationRequest"
                    }
                },
                "auto_create_next": {
                    "type": "boolean"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "entity.ExpenseSource": {
            "type": "string",
            "enum": [
                "manual",
                "sms",
                "whatsapp",
                "bank_api",
                "notification"
            ],
            "x-enum-comments": {
                "ExpenseSourceBankAPI": "API bancaria",
                "ExpenseSourceManual": "Entrada manual del usuario",
                "ExpenseSourceNotification": "Notificación del sistema",
                "ExpenseSourceSMS": "Capturado desde SMS bancario",
                "ExpenseSourceWhatsApp": "Enviado por WhatsApp"
            },
            "x-enum-descriptions": [
                "Entrada manual del usuario",
                "Capturado desde SMS bancario",
                "Enviado por WhatsApp",
                "API bancaria",
                "Notificación del sistema"
            ],
            "x-enum-varnames": [
                "ExpenseSourceManual",
                "ExpenseSourceSMS",
                "ExpenseSourceWhatsApp",
                "ExpenseSourceBankAPI",
                "ExpenseSourceNotification"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token para autenticación",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}